+++ TodoAPI/Repositories/ITodoRepository.cs
@@ -6,11 +6,11 @@ namespace TodoAPI.Repositories
 {
     public interface ITodoRepository
     {
-        Task<Todo> GetByIdAsync(long id);
+        Task<Todo> GetByIdAsync(int id);
         Task<List<Todo>> GetAllAsync();
         Task AddAsync(Todo todo);
         Task<bool> UpdateAsync(Todo todo);
-        Task<bool> DeleteAsync(long id);
-        Task<bool> ExistsAsync(long id);
+        Task<bool> DeleteAsync(int id);
+        Task<bool> ExistsAsync(int id);
     }
 }
